Implement the ability to view any account.
