+++ src/app/[lang]/events/page.tsx
@@ -18,6 +18,7 @@ import Link from "next/link"
 export const showQRCodeMinAccessLevel = AccessLevel.OFFICER
 export const createEventMinAccessLevel = AccessLevel.OFFICER
 export const createNewsMinAccessLevel = AccessLevel.OFFICER
+export const viewAnyAccountMinAccessLevel = AccessLevel.OFFICER
 
 const entriesPerPage = 10
