+++ src/localization/dictionaries/de.json
@@ -61,8 +61,13 @@
     "account_attendance_title": "Anwesenheitshistorie",
     "account_attendance_points": ":points Anwesenheitspunkte",
     "account_attendance_empty": "Du hast keine Veranstaltungen besucht.",
+    "account_any_attendance_empty": "",
     "account_spring_event": "Frühling :year",
     "account_fall_event": "Herbst :year",
+    "access_level_non_member": "",
+    "access_level_member": "",
+    "access_level_officer": "",
+    "access_level_advisor": "",
     "language_english": "English",
     "language_english_local": "Englisch",
     "language_german": "Deutsch",
