+++ src/localization/dictionaries/en.json
@@ -61,8 +61,13 @@
     "account_attendance_title": "Attendance History",
     "account_attendance_points": ":points attendance points",
     "account_attendance_empty": "You have not attended any events.",
+    "account_any_attendance_empty": "This user has not attendend any events.",
     "account_spring_event": "Spring :year",
     "account_fall_event": "Fall :year",
+    "access_level_non_member": "Guest",
+    "access_level_member": "Member",
+    "access_level_officer": "Officer",
+    "access_level_advisor": "Advisor",
     "language_english": "English",
     "language_english_local": "English",
     "language_german": "Deutsch",
