+++ src/localization/dictionaries/fr.json
@@ -61,11 +61,16 @@
     "account_attendance_title": "Historique de présence",
     "account_attendance_points": ":points points de présence",
     "account_attendance_empty": "Vous n'avez participé à aucun événement.",
+    "account_any_attendance_empty": "",
     "account_spring_event": "Printemps :year",
     "account_fall_event": "Automne :year",
-    "language_english": "Anglais (Anglais)",
+    "access_level_non_member": "",
+    "access_level_member": "",
+    "access_level_officer": "",
+    "access_level_advisor": "",
+    "language_english": "English",
     "language_english_local": "Anglais",
-    "language_german": "Deutsch (Allemand)",
+    "language_german": "Deutsch",
     "language_german_local": "Allemand",
     "language_french": "Français",
     "language_french_local": "Français",
