+++ src/localization/dictionaries/template.json
@@ -61,8 +61,13 @@
     "account_attendance_title": "",
     "account_attendance_points": "",
     "account_attendance_empty": "",
+    "account_any_attendance_empty": "",
     "account_spring_event": "",
     "account_fall_event": "",
+    "access_level_non_member": "",
+    "access_level_member": "",
+    "access_level_officer": "",
+    "access_level_advisor": "",
     "language_english": "",
     "language_german": "",
     "language_french": "",
